move replaceModalTitles.js from public to js subdirectory
