+++ public/js/replaceModalTitles.js
